requirements.txt
aiohttp==3.8.2
motor==3.0.0
uvloop==0.16.0
gunicorn==20.1.0
python-magic==0.4.26